First step in some refactoring of the NumberFormat; separate the cell numberformat properties from the actually code used to format a value, leavig just a callthrough stub
